requirements.txt
slack-sdk==3.23.0
spotipy==2.22.1
python-dotenv==1.0.0
youtube-dl==2021.12.17
yt-dlp==2023.11.16
requests==2.31.0
flask==2.3.3 